+++ lib/views/post_screen.dart
@@ -120,26 +120,3 @@ class PostScreen extends StatelessWidget {
         ));
   }
 }
-
-class Badge extends StatelessWidget {
-  final String level;
-  final String image;
-
-  Badge({@required this.level, @required this.image});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Text(
-          level,
-          style: TextStyle(color: kSubTextColor, fontSize: 13),
-        ),
-        Image.asset(
-          'assets/images/$image.png',
-          scale: 4,
-        ),
-      ],
-    );
-  }
-}
